feat: 초기화면 피팅룸으로 설정하고 sample모델 버튼 삭제

+++ app/src/main/java/com/example/lookup/data/UserModel.kt
@@ -33,12 +33,12 @@ class UserModel() {
         sizeInfo = SizeInfo(keyPoints!!, pixelHeight, height)
     }
 
-    fun getHeight():Float{
-        return height
+    fun getHeight():Int{
+        return height.toInt()
     }
 
-    fun getWeight():Float{
-        return weight
+    fun getWeight():Int{
+        return weight.toInt()
     }
 
     fun getSize():SizeInfo{
